Back/microservicios/task_service/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.0.5
PyJWT==2.10.1
psycopg2-binary==2.9.10
flask-cors
